% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik.R
\name{AICc}
\alias{AICc}
\title{Computes posterior sample of the pointwise corrected AIC method from a vartan object}
\usage{
AICc(obj)
}
\arguments{
\item{obj}{A varstan object of the time seires fitted model.}
}
\value{
A numeric array  of size R, containing the posterior samples of the AICc
for a varstan object, where R is the number of iterations. If multiple chains are
fitted, then the array is of length M*R, where m is the number of chains
}
\description{
Convenience function for computing the pointwise corrected  Akaike Information Criteria
method from a varstan object.
}
\examples{

\dontrun{
library(birth)
model = Sarima(birth,order = c(0,1,2),seasonal = c(1,1,1))
fit1 = varstan(model,chains = 1)

aic1 <- aic(fit1)
mean(aic1)
}

}
\author{
Asael ALonzo Matamoros
}
