% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima}
\alias{arima}
\title{Constructor arima(p,d,q) object}
\usage{
arima(ts,p,d,q)
}
\arguments{
\item{ts}{an multivariate time series}

\item{p}{an integer with the order of the ar(p) part}

\item{d}{an integer with the order of the diference (d)}

\item{q}{an integer with the order of the ma(q) part}

\item{sd:}{a list with the mgarch parameters}
}
\value{
a list with the components
\itemize{
 \item n: the length of the time series
 \item p: an integer with the order of the ar coefficients
 \item d: an integert with the order of diferences
 \item q: an integer with the order of the ma coefficients
 \item y: vector with the multivariate time series
 \item prior_ar: a matrix with the hyper-parameters for the var coeff icient
 \item prior_ma: a matrix with the hyper-parameters for the ma  coefficients
}
}
\description{
Constructor of the arima(p,d,q) object for bayesian estimation in STAN
}
\details{
The function returns  a list with the data for running stan() function of
 rstan package
}
\seealso{
\code{\link{garch}}
}
\author{
Asael Alonzo Matamoros
}
