% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garch.R
\name{arma}
\alias{arma}
\title{Adds an arma(p,q) object to a garch model}
\usage{
arma(p,q)
}
\arguments{
\item{p}{an integer with the order of the ar(p) part}

\item{q}{an integer with the order of the ma(q) part}
}
\description{
Adds an arma(p,q) object to a garch model
}
\details{
The function returns  a list with the data for running stan() function of
 rstan package
}
\author{
Asael Alonzo Matamoros
}
