% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{arma_check}
\alias{arma_check}
\title{Check the arma coefficients}
\usage{
arma_check(x)
}
\arguments{
\item{x}{is a vector with the hyper-parameter coefficients}
}
\value{
a vector with the checked values
}
\description{
checks for values in between -1 and 1 in a real vector
}
\author{
Asael Alonzo Matamoros
}
