% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varstan.R
\name{as.stan}
\alias{as.stan}
\title{Get the stanfit object generated by the rstan package}
\usage{
as.stan(obj)
}
\arguments{
\item{obj}{a varstan object}
}
\value{
a stanfit object
}
\description{
Get the stanfit object generated by the rstan package
}
\author{
Asael Alonzo Matamoros
}
