% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.varstan}
\alias{autoplot.varstan}
\title{autoplot methods for varstan models}
\usage{
\method{autoplot}{varstan}(object, par = "fit", probs = 0.9, ...)
}
\arguments{
\item{par}{a character string with the desired plot. Valid values are \code{"fit"} for the
model's fitted values plot, \code{"residuals"} for the residuals posterior mean plot,
and \code{"parameters"} for density and trace plots for the sampled model parameter.
The default value is \code{"fit"}.}

\item{...}{Further arguments passed to  \code{mcmc_combo}.}

\item{x}{An object of class \code{varstan}.}
}
\value{
An autoplot object from ggplot2 class.
}
\description{
Preliminar autoplot methods for varstan models only valid for univariate time series models.
The function prints the fitted values time series, the trace and density plots for the
sampled model parameters, or the residuals' posterior mean time series.
}
\examples{
\dontrun{
 sf1 = auto.sarima(ts = birth)
 # fitted model
 autoplot(sf1)

 # residuals
 autoplot(sf1,type = "residuals)

 # parameters
 autoplot(sf1,type = "parameter)
}

}
