% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bekk.R
\name{Bekk}
\alias{Bekk}
\title{Constructor VARMA(p,q)-Bekk(s,k) object.}
\usage{
Bekk(ts,order = c(1,1,0),varma = c(0,0),genT = FALSE,series.name = NULL)
}
\arguments{
\item{ts}{an multivariate time series.}

\item{order}{A numeric vector for specify the Bekk  part for variance, same as
order parameter:  the three components (s, k, h) are the arch order,
the garch order and the garch for the mean (mgarch) order.}

\item{varma}{an optional numeric vector for specification of a VARMA model: the two
components (p, q) are the VAR order, the number of differences, and the VMA order.}

\item{genT}{a boolean value to specify a Generalized a t-student model Cruz (2015).}

\item{series.name}{an optional string vector with the time series names.}
}
\description{
Constructor of the VARMA(p,q)-Bekk(s,k) object for Bayesian estimation in STAN.
}
\details{
The function returns  a list with the data for running \code{stan()} function of
 \pkg{rstan} package


If a Bekk model is specified, then sigma0 represents the triangular inferior matrix
alpha0 of a Bekk model.

The default priors used in VARMA(p,q)-MBekk(s,k,h) model are:

\itemize{
 \item{"ar"}{ar ~ normal(0,0.5)}
 \item{"ma"}{ma ~ normal(0,0.5)}
 \item{"mu0"}{mu0 ~ normal(0,1)}
 \item{"sigma0"}{sigma0 ~ t-student(0,1,7)}
 \item{"arch"}{arch ~ normal(0,0.5)}
 \item{"garch"}{garch ~ normal(0,0.5)}
 \item{"mgarch"}{mgarch ~ normal(0,0.5)}
 \item{"dfv"}{dfv ~ gamma(2,0.1)}
}

For changing the default prior use the \code{set_prior()} function.
}
\examples{
# Declare a Bekk model for the Astrovan data
model = Bekk(Astrovan,order = c(1,1,0),varma = c(1,0))
model


}
\references{
Polasek, W. (2000). A Multivariate GARCH-M Model for Exchange Rates in the US,
Germany and Japan. \emph{Springer Berlin Heidelberg}. 355-363.
\code{doi: 10.1007/978-3-642-57280-7_39}

Fioruci, J. and Ehlers, R. and Andrade, M. (2014). Bayesian multivariate GARCH
models with dynamic correlations and asymmetric error distributions.
\emph{Journal of Applied Statistics}. 41(2), 320 - 331.

Berg, T. O. (2016) Multivariate Forecasting with BVARs and DSGE Models.
\emph{J. Forecast}. 35(1), 718- 740.
\code{doi: 10.1002/for.2406}.
}
\seealso{
\code{\link{garch}} \code{\link{Sarima}} \code{\link{varma}}
}
\author{
Asael Alonzo Matamoros
}
