% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{check_loc}
\alias{check_loc}
\title{Check if the value is in the domain of a location parameter
 @export}
\usage{
check_loc(x)
}
\description{
Check if the value is in the domain of a location parameter
 @export
}
