% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{complete}
\alias{complete}
\title{Complete vector}
\usage{
complete(x,d,x0)
}
\arguments{
\item{x}{is a vector with the hyper-parameter coefficients}

\item{d}{an integer with the vector dimention}

\item{x0}{a real with the preliminar value for the hyper-parameter}
}
\value{
a vector with the hyper-parameter value
}
\description{
Completes the prior vector hiper parameters to their default value
}
\author{
Asael Alonzo Matamoros
}
