% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{dif}
\alias{dif}
\title{Extract the initial values of a differenced time series}
\usage{
dif(ts, d = 0, D = 0, period = 1)
}
\description{
@param ts An univariate time series
 @param d the number of differences
 @param D the number of seasonal differences
 @param period the time series period
}
\details{
@author Asael Alonzo Matamoros

 @return A list with the differenced time series, the initials values
 for inverse differences

 @noRd
}
