% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Others.R
\name{distribution}
\alias{distribution}
\title{Print the  available prior distribution for the model parameters}
\usage{
distribution(par)
}
\arguments{
\item{par}{a string value with the desired parameter, if not know the parameter
use the function \code{parameters} with the model class.}
}
\value{
a string with the prior distributions for the desired parameter
}
\description{
The functions prints the available distributions for the model
parameters in \pkg{varstan}.
}
\examples{
 version()
 parameters(classes = "Sarima")
 distribution(par = "ar")

}
\seealso{
\code{parameters}, \code{version}
}
\author{
Asael Alonzo Matamoros
}
