% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{extract_estimate}
\alias{extract_estimate}
\title{point estimate of univariate parameters}
\usage{
point_estimate.arima(model,fit)
}
\arguments{
\item{fit:}{a stanfit object}

\item{model:}{a varbekk model}

\item{par:}{the desired parameter}

\item{robust:}{a boolean for obtain the robust estimation}
}
\value{
a vector with all the elected parameters
}
\description{
get the point estimate of univariate parameters in a STAN model
}
\details{
The function returns a vector with the fitted parameters
}
\author{
Asael Alonzo Matamoros
}
