% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{extract_mts}
\alias{extract_mts}
\title{Extract estimate of multivariate time series parameters from a stanfit object}
\usage{
extract_mts(obj, par, d = 1, lag = 1, robust = FALSE)
}
\description{
@param obj: varstan object
 @param par: a string with the desired parameter
 @param d: the dimension of the time series
 @param robust: a boolean value ofr robust estimate
 @param lag: the max lags desired
}
\details{
@author Asael Alonzo Matamoros

@return
}
