% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{extract_ts}
\alias{extract_ts}
\title{Extract estimate time series parameters from a stanfit object}
\usage{
extract_ts(obj, par, lag = 1, robust = FALSE)
}
\description{
@param obj: varstan object
 @param par: a string with the desired parameter
 @param robust: a boolean value ofr robust estimate
 @param lag: the max lags desired
}
\details{
@author Asael Alonzo Matamoros

 @return
}
