% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{fit.arima}
\alias{fit.arima}
\title{Fit an arima model}
\usage{
fit.arima(model,chains,iter,warmup,...)
}
\arguments{
\item{model}{A time series object for the varstan models}

\item{chains}{the number of chains to be run}

\item{iter}{the number of iteration per chain}

\item{warmup}{the number of initial iteration to be burned}

\item{adapt.delta}{the thin of the jumps in a HMC method}
}
\value{
a stanfit object
}
\description{
Fit a arima(p,d,q) model  in STAN
}
\details{
The function returns a list with the fitted model in stan
}
\author{
Asael Alonzo Matamoros
}
