% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{fit_garch}
\alias{fit_garch}
\title{Fit a mgarch model}
\usage{
fit.garch(model,chains,iter,warmup,...)
}
\arguments{
\item{model}{A time series object for the varstan models}

\item{chains}{the number of chains to be run}

\item{iter}{the number of iteration per chain}

\item{warmup}{the number of initial iteration to be burned}

\item{adapt.delta}{the thin of the jumps in a HMC method}

\item{tree.depth}{maximum  depth of the trees  evaluated
during each iteration}
}
\value{
a stanfit object
}
\description{
Fit a garch(s,k,h) model  in STAN
}
\details{
The function returns a list with the fitted model in stan
}
\author{
Asael Alonzo Matamoros
}
