% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{fit_varma}
\alias{fit_varma}
\title{Fit a  varma model}
\usage{
fit.varma(model)
}
\arguments{
\item{model}{A time series object for the varstan models}

\item{chains}{the number of chains to be run}

\item{iter}{the number of iteration per chain}

\item{warmup}{the number of initial iteration to be burned}

\item{adapt.delta}{the thin of the jumps in a HMC method}

\item{t.student}{a boolean if the model has generalized t-student distribution}
}
\value{
a stanfit object
}
\description{
Fit a var(p)-ma(q) model  in STAN
}
\details{
The function returns a list with the fitted model in stan
}
\author{
Asael Alonzo Matamoros
}
