% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garch.R
\name{garch}
\alias{garch}
\title{Constructor garch(s,k,h) object}
\usage{
garch(ts,s,k)
}
\arguments{
\item{ts}{an univariateivariate time series}

\item{s}{an integer with the order of the arch(s) part}

\item{k}{an integer with the order of the garch(k) part}

\item{h}{an integer with the order of the mean part}

\item{mu0:}{a list with the mgarch parameters}
}
\value{
a list with the components
\itemize{
 \item n: the length of the time series
 \item p: an integer with the order of the ar coefficients
 \item d: an integert with the order of diferences
 \item q: an integer with the order of the ma coefficients
 \item y: vector with the multivariate time series
 \item prior_ar: a matrix with the hyper-parameters for the var coeff icient
 \item prior_ma: a matrix with the hyper-parameters for the ma  coefficients
}
}
\description{
Constructor of the garch(s,k,h) object for bayesian estimation in STAN
}
\details{
The function returns  a list with the data for running stan() function of
 rstan package
}
\seealso{
\code{\link{arima}}
}
\author{
Asael Alonzo Matamoros
}
