% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/Misc.R
\name{get_df}
\alias{get_df}
\title{Set the generic function for fit in a varstan class}
\usage{
get_df(fit, model, robust, ...)
}
\description{
Rstan fit function
}
\details{
@method get_df

@export
=======
% Please edit documentation in R/varstan.R
\name{get_df}
\alias{get_df}
\title{Get the degree freedom values of a Generalized t-student varma model}
\usage{
get_df.varma(fit)
}
\arguments{
\item{obj:}{a varstan object}

\item{robust:}{a boolean for obtain the robust estimation}
}
\value{
a data frame with all the important fitted parameters
}
\description{
get the degree freedom values of a Generalized t-student varma(p,q) model
}
\details{
The function returns a data.frame object with the degree freedom values
}
\author{
Asael Alonzo Matamoros
>>>>>>> s3 methods corrected
}
