% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{get_fit}
\alias{get_fit}
\title{Set the generic function for get fitted values in a varstan class}
\usage{
get_fit(model, fit, robust, ...)
}
\description{
A point estimate of a desired model
}
\details{
@method get_fit

@export
}
