% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varstan.R
\name{get_fit.vastan}
\alias{get_fit.vastan}
\title{Get the fitted values of the fitted model}
\usage{
point_estimate.varma(fit)
}
\arguments{
\item{obj:}{a varstan object}

\item{robust:}{a boolean for obtain the robust estimation}

\item{conf:}{a value between 0 and 1 with the desired confidence (not for
varma models)}
}
\value{
a data frame with all the important fitted parameters
}
\description{
The function returns a matrix with the fitted values
}
\author{
Asael Alonzo Matamoros
}
