% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_fit.R
\name{get_fit_arima}
\alias{get_fit_arima}
\title{Get the fitted values of an arima model}
\usage{
point_estimate.arima(model,fit)
}
\arguments{
\item{fit:}{a stanfit object}

\item{model:}{the arima model}

\item{robust:}{a boolean for obtain the robust estimation}
}
\value{
a data frame with all the important fitted parameters
}
\description{
get the fitted values of an arima(p,d,q) model  in STAN
}
\details{
The function returns a data.frame object with the fitted values
}
\author{
Asael Alonzo Matamoros
}
