% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_values.R
\name{get_fit_garch}
\alias{get_fit_garch}
\title{Get the fitted values of an garch model}
\usage{
point_estimate.garch(model,fit)
}
\arguments{
\item{fit:}{a stanfit object}

\item{model:}{the arima model}

\item{robust:}{a boolean for obtain the robust estimation}
}
\value{
a data frame with all the important fitted parameters
}
\description{
get the fitted values of an garch(s,k,h) model  in STAN
}
\details{
The function returns a data.frame object with the fitted values
}
\author{
Asael Alonzo Matamoros
}
