% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_values.R
\name{get_fit_varma}
\alias{get_fit_varma}
\title{Get the fitted values of an varma model}
\usage{
point_estimate.varma(fit)
}
\arguments{
\item{model:}{a varma model object}

\item{fit:}{a stanfit object}

\item{robust:}{a boolean for obtain the robust estimation}

\item{conf:}{a value between 0 and 1 with the desired confidence}
}
\value{
a data frame with all the important fitted parameters
}
\description{
get the fitted values of an varma(p,q) model  in STAN
}
\details{
The function returns a data.frame object with the fitted values
}
\author{
Asael Alonzo Matamoros
}
