% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varma.R
\name{get_lag_varma}
\alias{get_lag_varma}
\title{Get the lag parameters of a varma model}
\usage{
get_lag.varma(type,model,fit,robust)
}
\arguments{
\item{type:}{the parameter to be extracted}

\item{fit:}{a stanfit object}

\item{model:}{a model object}

\item{robust:}{a boolean for obtain the robust estimation}
}
\value{
a data frame with all the important fitted parameters

a data frame with all the important fitted parameters
}
\description{
get the degree freedom values of a tvarma(p,q) model  in STAN
}
\details{
The function returns a data.frame object with the degree freedom values
}
\author{
Asael Alonzo Matamoros
}
