% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{get_one_prior_ma}
\alias{get_one_prior_ma}
\title{Get one Prior distribution for a ma model}
\usage{
get_one_prior(dat,lag = 1)
}
\arguments{
\item{dat:}{an arima model speficied in varstan.}

\item{lag:}{the lag of the parameter which the prior is defined if lag = 0, then the prior will be applied
for all the lags.}
}
\description{
Get one Prior distribution for a ma model
}
\author{
Asael Alonzo Matamoros
}
