% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varstan.R
\name{get_order}
\alias{get_order}
\title{Extracts all the order coeffients in a list}
\usage{
get_order(obj)
}
\arguments{
\item{obj:}{A varstan object}
}
\description{
Extracts all the order coeffients in a list
}
