% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Excluded and included parameters in a defined time series model}
\usage{
get_params(obj, ...)
}
\arguments{
\item{obj}{a varstan object}
}
\description{
Excluded and included parameters in a defined time series model
}
