% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{get_prior}
\alias{get_prior}
\title{Generic function for setting a prior to an specify parameter}
\usage{
set_prior(type,lag,par1,par2,dist)
}
\arguments{
\item{dat:}{an arima model speficied in varstan.}

\item{type:}{the type of parameter wich a prior is defined could be mu, sigma0, ar, ma, arch, garch, mgarch}

\item{lag:}{the lag of the parameter which the prior is defined if lag = 0, then the prior will be applied
for all the lags.}
}
\description{
Generic function for setting a prior to an specify parameter
}
\author{
Asael Alonzo Matamoros
}
