% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{get_prior_arch}
\alias{get_prior_arch}
\title{Get Prior distribution for an arch model}
\usage{
get_prior(dat,lag)
}
\arguments{
\item{dat:}{an arima model speficied in varstan.}

\item{lag:}{the lag of the parameter which the prior is defined if lag = 0, then the prior will be applied
for all the lags.}
}
\description{
Get Prior distribution for an arch model
}
\author{
Asael Alonzo Matamoros
}
