% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{get_prior_dfv}
\alias{get_prior_dfv}
\title{Set Prior distribution for a degree gfreedom v parameter}
\usage{
get_prior_dfv(dat)
}
\arguments{
\item{dat:}{an varma model speficied in varstan.}
}
\description{
Set a prior distribution for a degree freedom v parameter
 in a Generalized t-student varma model
}
\author{
Asael Alonzo Matamoros
}
