% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{get_prior_mu0}
\alias{get_prior_mu0}
\title{Get Prior distribution for a mu parameter}
\usage{
get_prior_mu0(dat)
}
\arguments{
\item{dat:}{an arima model speficied in varstan.}
}
\description{
Generic function for setting a prior to the mean parameters
}
\author{
Asael Alonzo Matamoros
}
