% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{get_residuals}
\alias{get_residuals}
\title{Set the generic function for get_residuals in a varstan class}
\usage{
get_residuals(model, fit, robust, ...)
}
\description{
Rstan fit function
}
\details{
@method get_residuals

@export
}
