% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varstan.R
\name{get_residuals.vastan}
\alias{get_residuals.vastan}
\title{Get the residual values of the fitted model}
\usage{
get_residuals(obj)
}
\arguments{
\item{obj:}{a varstan object}

\item{robust:}{a boolean for obtain the robust estimation}
}
\value{
a data frame with all the important fitted parameters
}
\description{
The function returns a matrix with the residual values
}
\author{
Asael Alonzo Matamoros
}
