% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{inv_dif}
\alias{inv_dif}
\title{Inverse difference time series to forecasted values}
\usage{
inv_dif(ts, init, inits)
}
\description{
@param ts: An univariate time series
 @param init: the initial values for normal difference
 @param inits: the initial values for seasonal difference
}
\details{
@author Asael Alonzo Matamoros

 @return A vector with the inverse difference time series
}
