% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik.R
\name{log_lik.varstan}
\alias{log_lik.varstan}
\alias{log_lik}
\title{Extract posterior sample of the pointwise log-likelihood from a vartan object}
\usage{
\method{log_lik}{varstan}(obj, permuted = TRUE)
}
\arguments{
\item{obj}{A varstan object of the time seires fitted model.}

\item{permuted}{A logical scalar indicating whether the draws after 
the warmup period in each chain should be permuted and merged. 
If FALSE, the original order is kept. For each stanfit object, 
the permutation is fixed (i.e., extracting samples a second time 
will give the same sequence of iterations).}
}
\value{
Usually, an S x N matrix containing the pointwise log-likelihood
 samples, where S is the number of samples and N is the number 
 of observations in the data. If 
 \code{permuted} is \code{FALSE}, an S x N x R array is returned, 
 where R is the number of fitted chains.
}
\description{
Convenience function for extracting the pointwise log-likelihood
matrix or array from a fitted Stan model.
}
