% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik.R
\name{loo.varstan}
\alias{loo.varstan}
\alias{loo}
\title{Leave-one-out cross-validation}
\usage{
loo.varstan(obj)
}
\arguments{
\item{obj}{A varstan object}
}
\description{
The \code{loo} method for varstan objects. Computes approximate
leave-one-out cross-validation using Pareto smoothed importance sampling
(PSIS-LOO CV).
}
\examples{

\dontrun{
library(birth)
model = Sarima(birth,order = c(0,1,2),seasonal = c(1,1,1))
fit1 = varstan(model,chains = 1)

loo1 <- loo(fit1)
loo1
}
}
\seealso{
* The **loo** package [vignettes](https://mc-stan.org/loo/articles/index.html)
   for demonstrations.
 * [psis()] for the underlying Pareto Smoothed Importance Sampling (PSIS)
   procedure used in the LOO-CV approximation.
 * [pareto-k-diagnostic] for convenience functions for looking at diagnostics.
 * [loo_compare()] for model comparison.
}
