% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varma.R
\name{mbekk}
\alias{mbekk}
\title{Adds a Bekk(s,k,h) object to an arima model}
\usage{
bekk(s,k,h)
}
\arguments{
\item{s}{an integer with the order of the arch(s) part}

\item{k}{an integer with the order of the garch(k) part}

\item{h}{an integer with the order of the mgarch(h) part}
}
\description{
Adds a Bekk(s,k,h) object to an arima model
}
\details{
The function returns  a list with the data for running stan() function of
 rstan package
}
\author{
Asael Alonzo Matamoros
}
