% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive.R
\name{naive}
\alias{naive}
\title{naive and Random Walk Forecasts}
\usage{
naive(ts,seasonal = FALSE,m = 0)
}
\arguments{
\item{ts}{a numeric vector or time series of class \code{ts}}

\item{seasonal}{a boolean value for select a seasonal random walk instead}

\item{m}{an integer value}
}
\value{
An object of class naive.
}
\description{
naive is the model constructor for a random walk  model applied to \code{y}.
This is equivalent to an ARIMA(0,1,0) model. \code{naive()} is simply a wrapper
to  mantain forecast package similitude. \code{seasonal} returns the model constructor
for a seasonal random walkequivalent to an ARIMA(0,0,0)(0,1,0)m model where m is the
seasonal period.
}
\details{
The random walk with drift model is \deqn{Y_t=c + Y_{t-1} + Z_t}{Y[t]=c +
Y[t-1] + Z[t]} where \eqn{Z_t}{Z[t]} is a normal iid error. Forecasts are
given by \deqn{Y_n(h)=ch+Y_n}{Y[n+h]=ch+Y[n]}.

The seasonal naive model is \deqn{Y_t= Y_{t-m} + Z_t}{Y[t]=Y[t-m] + Z[t]}
where \eqn{Z_t}{Z[t]} is a normal iid error.
}
\examples{

model = naive(birth,seasonal = TRUE)
model

}
\seealso{
\code{\link{Sarima}}
}
\author{
Asael ALonzo Matamoros
}
