% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{no_negative_check}
\alias{no_negative_check}
\title{No negative values in vector}
\usage{
no_negative_check(x,x0)
}
\arguments{
\item{x}{is a vector with the hyper-parameter coefficients}

\item{x0}{a real value to replace in a non positive value}

\item{d}{an integer with the vector dimention}
}
\value{
a vector with the checked values
}
\description{
checks for no negative values in a vector (x_i >= 0 ) usefull for checking
lags in models
}
\author{
Asael Alonzo Matamoros
}
