% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Others.R
\name{parameters}
\alias{parameters}
\title{Print the parameters of a model class}
\usage{
parameters(classes = "Sarima")
}
\arguments{
\item{classes}{a string value with the desyred model class (Sarima,garch,varma,ect,...),
for knowing the current avaliable model classes use the \code{version()} function}
}
\value{
a string with the prior distributions for the desired parameter
}
\description{
The functions prints the parameters in one of the defined model classes in varstan
}
\examples{
 version()
 parameters(classes = "Sarima")
 distribution(par = "ar")

}
\seealso{
\code{parameters}, \code{version}
}
\author{
Asael Alonzo Matamoros
}
