% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{point_estimate.arima}
\alias{point_estimate.arima}
\title{point estimate of an arima model}
\usage{
point_estimate.arima(model,fit)
}
\arguments{
\item{fit:}{a stanfit object}

\item{model:}{a varbekk model}

\item{robust:}{a boolean for obtain the robust estimation}
}
\value{
a data frame with all the important fitted parameters
}
\description{
get the point estimate of an arima(p,d,q) model  in STAN
}
\details{
The function returns a data.frame object with the fitted parameters
}
\author{
Asael Alonzo Matamoros
}
