% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/Misc.R
\name{point_estimate}
\alias{point_estimate}
\title{Set the generic function for point_estimate a varstan class}
\usage{
point_estimate(model, fit, robust, conf, ...)
}
\description{
A point estimate of a desired model
}
\details{
@method point_estimate

@export
=======
% Please edit documentation in R/point_estimate.R
\name{point_estimate}
\alias{point_estimate}
\title{Point estimate of  a varstan object}
\usage{
point_estimate(obj)
}
\arguments{
\item{obj:}{a varstan object}

\item{robust:}{A boolean value, if its true it returns the posterior median}
}
\value{
a list with the components
\itemize{
 \item A matrix with the posterior value of the var coefficients
 \item A matrix with the posterior value of the scale parameter
 \item A vector with the posterior value of the degree freedom
 \item A time series with the posterior value of the correction parameter
 for the variance
}
}
\description{
point estimates of the model  in a varstan object
}
\author{
Asael Alonzo Matamoros
>>>>>>> s3 methods corrected
}
