% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varma.R
\name{point_estimate.varma}
\alias{point_estimate.varma}
\title{point estimate of an varma model}
\usage{
point_estimate.varma(model,fit)
}
\arguments{
\item{fit:}{a stanfit object}

\item{model:}{a varbekk model}

\item{robust:}{a boolean for obtain the robust estimation}

\item{par:}{the wanted parameters, by default the option all is used}
}
\value{
a data frame with all the important fitted parameters
}
\description{
Get the point estimate of an varma(p,q) model  in STAN
}
\details{
The function returns a data.frame object with the fitted parameters
}
\author{
Asael Alonzo Matamoros
}
