% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varstan.R
\name{point_estimate.varstan}
\alias{point_estimate.varstan}
\title{Point estimate of  a varstan object}
\usage{
point_estimate(obj)
}
\arguments{
\item{obj:}{a varstan object}

\item{robust:}{A boolean value, if its true it returns the posterior median}
}
\value{
a list with the components
\itemize{
 \item A matrix with the posterior value of the var coefficients
 \item A matrix with the posterior value of the scale parameter
 \item A vector with the posterior value of the degree freedom
 \item A time series with the posterior value of the correction parameter
 for the variance
}
}
\description{
point estimates of the model  in a varstan object
}
\author{
Asael Alonzo Matamoros
}
