% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_estimate.R
\name{point_estimate_garch}
\alias{point_estimate_garch}
\title{point estimate of an garch model}
\usage{
point_estimate.garch(model,fit)
}
\arguments{
\item{fit:}{a stanfit object}

\item{model:}{a varbekk model}

\item{robust:}{a boolean for obtain the robust estimation}
}
\value{
a data frame with all the important fitted parameters
}
\description{
get the point estimate of an garch(s,k,h) model  in STAN
}
\details{
The function returns a data.frame object with the fitted parameters
}
\author{
Asael Alonzo Matamoros
}
