% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{positive_check}
\alias{positive_check}
\title{Positive values in vector}
\usage{
positive_check(x,x0)
}
\arguments{
\item{x}{is a vector with the hyper-parameter coefficients}

\item{x0}{a real value to replace in a non positive value}

\item{d}{an integer with the vector dimention}
}
\value{
a vector with the checked values
}
\description{
checks for positive values in a vector (x_i > 0) usefull for
checking scale parameters
}
\author{
Asael Alonzo Matamoros
}
