% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_estimate.R
\name{posterior_estimate}
\alias{posterior_estimate}
\title{Generic function and method for the posteriorpoint estimate of all  parameters in varstan object}
\usage{
point_estimate(obj)
}
\arguments{
\item{obj:}{a varstan object}

\item{robust:}{A boolean value, if its \code{TRUE} it returns the median of the posterior distribution,
And if its \code{FALSE} it returns the mean, by default is the \code{FALSE} value}
}
\value{
a list with the components of mu0, sigma0, ar, ma, sar, sma, arch, garch, mgarch, and breg,
of one of the defined varstan models
}
\description{
It returns a list with the point estimate value of the posterior distribution
of \strong{all} the model parameters of the varstan object. By defaults it returns the
posterior mean.
}
\author{
Asael Alonzo Matamoros
}
