% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_estimate.R
\name{posterior_estimate.varstan}
\alias{posterior_estimate.varstan}
\alias{posterior_estimate}
\title{Generic function and method for the posterior point estimate of all  parameters in varstan object}
\usage{
posterior_estimate(object,robust = FALSE,...)

posterior_estimate(object, robust = FALSE, ...)
}
\arguments{
\item{object}{a varstan object}

\item{robust}{A boolean value, if its \code{TRUE} it returns the median of the posterior distribution,
And if its \code{FALSE} it returns the mean, by default is the \code{FALSE} value}

\item{...}{Further arguments passed to  \code{posterior_estimate}.}
}
\value{
a list with the components of mu0, sigma0, ar, ma, sar, sma, arch, garch, mgarch, and breg,
of one of the defined varstan models
}
\description{
It returns a list with the point estimate value of the posterior distribution
of \strong{all} the model parameters of the varstan object. By defaults it returns the
posterior mean.
}
\author{
Asael Alonzo Matamoros
}
