% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predict.R
\name{posterior_predict_garch}
\alias{posterior_predict_garch}
\title{Draw from posterior predictive distribution of an arma-garch model}
\usage{
posterior_predict_garch(obj)
}
\arguments{
\item{obj:}{a varstan object}

\item{h:}{An integer indicating the number of predictions. The default number
of predictions is 1.}

\item{robust:}{A boolean for obtain the robust estimation. The default}

\item{draws:}{An integer indicating the number of draws to return. The default
number of draws is 1000}

\item{seed:}{An optional \code{\link[=set.seed]{seed}} to use.}
}
\value{
A \code{draws} by \code{h} data.frame of simulations from the
  posterior predictive distribution. Each row of the data.frame is a vector of
  predictions generated using a single draw of the model parameters from the
  posterior distribution.
}
\description{
Draw from posterior predictive distribution of an arma-garch model
}
\author{
Asael Alonzo Matamoros
}
