% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_residuals.R
\name{posterior_residuals.varstan}
\alias{posterior_residuals.varstan}
\alias{posterior_residuals}
\title{Generic function and method for extract the residual of a varstan object}
\usage{
posterior_residuals(object,robust = FALSE,...)

posterior_residuals(object, robust = FALSE, ...)
}
\arguments{
\item{object}{A varstan object, \code{\link[=varstan]{varstan}}}

\item{robust}{A boolean value, if its \code{TRUE} it returns the median of the posterior distribution,
and if its \code{FALSE} it returns the mean, by default is the \code{FALSE} value}

\item{...}{Further arguments passed to  \code{posterior_residual}.}
}
\value{
An array with the posterior estimate of the residuals of the time series model,
if a varma model is used, then the return will be a matrix with columns as the
dimension of the time series, and rows as the length
}
\description{
The function returns the posterior estimate of the residuals
of a varstan model, similar to the residual functions of other
packages.
}
\details{
This function only extracts the pointwise estimate of the time series resiudals
for extracting all the data use extract_stan or posterior_intervals function
}
\author{
Asael Alonzo Matamoros
}
