% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Bekk}
\alias{print.Bekk}
\title{Print a Bekk model}
\usage{
\method{print}{Bekk}(x, ...)
}
\arguments{
\item{x}{a Bekk model from the varstan package}

\item{...}{additional values need in print methods}
}
\description{
Print a Bekk model
}
