% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Sarima}
\alias{print.Sarima}
\title{Print a Sarima model}
\usage{
\method{print}{Sarima}(obj, ...)
}
\arguments{
\item{obj}{a Sarima model from the varstan package}
}
\description{
Print a Sarima model
}
