% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SVM}
\alias{print.SVM}
\title{Print a Stochastic Volatility model}
\usage{
\method{print}{SVM}(obj, ...)
}
\arguments{
\item{obj}{a SVM model from the varstan package}
}
\description{
Print a Stochastic Volatility model
}
