% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{repeat_value}
\alias{repeat_value}
\title{Replicate Elements of Vector}
\usage{
repeat_value(x,d,x0)
}
\arguments{
\item{x}{is a vector with the hyper-parameter coefficients}

\item{d}{an integer with the vector dimention}

\item{x0}{a real value to replace in a non positive value}
}
\value{
a vector with repeated values
}
\description{
replicate a value d times if the value is different than x0
}
\author{
Asael Alonzo Matamoros
}
