% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{report.arima}
\alias{report.arima}
\title{Print a report of the constructed model}
\usage{
\method{report}{arima}(dat)
}
\arguments{
\item{dat:}{an arima object}
}
\description{
Print a report of the constructed model
}
