% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garch.R
\name{report.garch}
\alias{report.garch}
\title{Print a report of the constructed model}
\usage{
\method{report}{garch}(dat)
}
\arguments{
\item{dat:}{a garch object}
}
\description{
Print a report of the constructed model
}
