% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/Misc.R
\name{report}
\alias{report}
\title{Set the generic function for report a varstan class}
\usage{
report(dat, ...)
}
\description{
@export
}
\details{
@method report

@S3method report
=======
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Print a report of the constructed model}
\usage{
report(dat)
}
\arguments{
\item{dat:}{an arima object}
}
\description{
Print a report of the constructed model
>>>>>>> s3 methods corrected
}
