% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varma.R
\name{report.varma}
\alias{report.varma}
\title{Print a report of the constructed model}
\usage{
\method{report}{varma}(dat)
}
\arguments{
\item{dat:}{an varma object}
}
\description{
Print a report of the constructed model
}
