% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sarima.R
\name{Sarima}
\alias{Sarima}
\title{Constructor Seasonal arima model}
\usage{
Sarima(ts,order = c(0,0,0),seasonal = c(0,0,0),xreg = NULL,period = 0)
}
\arguments{
\item{ts}{a univariate time series}

\item{order:}{A specification of the non-seasonal part of the ARIMA model: the
three components (p, d, q) are the AR order, the degree of differencing, and the
MA order.}

\item{seasonal:}{A specification of the seasonal part of the ARIMA model,same as
order parameter:  the three components (p, d, q) are the seasonal AR order,
the degree of seasonal differencing, and the seasonal MA order.}

\item{xreg:}{Optionally, a numerical matrix of external regressors,
which must have the same number of rows as ts. It should not be a data frame.}

\item{period:}{an integer specifying the periodicity of the time series by
deafault the value frequency(ts) is used.}
}
\description{
Constructor of the Sarima model  for bayesian estimation in STAN
}
\details{

}
\examples{
# Declare a seasonal arima model for the birth data of asta package

model = Sarima(bth,order = c(0,1,2),seasonal = c(1,1,1))
model


}
\references{
Box, George; Jenkins, Gwilym (1970).
 Time Series Analysis: Forecasting and Control. San Francisco: Holden-Day.
}
\seealso{
\code{\link{garch}} \code{\link{auto.arima}} \code{\link{set_prior}}
}
\author{
Asael Alonzo Matamoros
}
