% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{set_prior_dfv}
\alias{set_prior_dfv}
\title{Set Prior distribution for a degree gfreedom v parameter}
\usage{
set_prior(dat,par1,par2,dist,df)
}
\arguments{
\item{par1}{The first hyper-parameter of the prior distribution}

\item{par2}{The second hyper-parameter of the prior distribution}

\item{dat:}{a varma model speficied in varstan.}

\item{dist:}{the dist of the prior distribution, could be: a normal,beta, and uniform}

\item{df:}{The parameter for the degree freedom only for t-student and chi-square}
}
\description{
Set a prior distribution for a degree freedom v parameter
 in a Generalized t-student varma model
}
\author{
Asael Alonzo Matamoros
}
