% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{set_prior_mgarch}
\alias{set_prior_mgarch}
\title{Set Prior distribution for a mgarch model}
\usage{
set_prior(type,lag,par1,par2,dist)
}
\arguments{
\item{par1}{The first hyper-parameter of the prior distribution}

\item{par2}{The second hyper-parameter of the prior distribution}

\item{dat:}{a garch model speficied in varstan.}

\item{lag:}{the lag of the parameter which the prior is defined if lag = 0, then the prior will be applied
for all the lags.}

\item{dist:}{the distribution of the prior distribution, could be: a normal,beta, and uniform}
}
\description{
Generic function for setting a prior to garch parameters
}
\author{
Asael Alonzo Matamoros
}
