% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{summary}
\alias{summary}
\title{Set the generic function for summary a varstan class}
\usage{
summary(model, fit, robust, conf, ...)
}
\description{
A report of the desired model
}
\details{
@method summary

@export
}
