% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varma.R
\name{summary.varma}
\alias{summary.varma}
\title{Summary of  a varma model}
\usage{
summary.varma(model,fit)
}
\arguments{
\item{fit:}{a stanfit object}

\item{model:}{a varbekk model}

\item{robust:}{a boolean for obtain the robust estimation}

\item{conf:}{a value between 0 and 1 with the desired confidence}
}
\value{
a data frame with all the important fitted parameters
}
\description{
Summary of an varma(p,q) model  in STAN
}
\details{
The function returns a data.frame object with the fitted parameters
}
\author{
Asael Alonzo Matamoros
}
