% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varstan.R
\name{summary.varstan}
\alias{summary.varstan}
\title{Summary of  a varstan object}
\usage{
summary(obj)
}
\arguments{
\item{obj:}{a varstan object}

\item{robust:}{a boolean for obtain the robust estimation}

\item{conf:}{a value between 0 and 1 with the desired probability in the
credible intervals}
}
\value{
a list with the components
\itemize{
 \item summary with all the important fitted parameters
}
}
\description{
Summary of the model estimates in a varstan object
}
\author{
Asael Alonzo Matamoros
}
