% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_arima}
\alias{summary_arima}
\title{Summary of  an arima model}
\usage{
summary.arima(model,fit)
}
\arguments{
\item{fit:}{a stanfit object}

\item{model:}{a varbekk model}

\item{robust:}{a boolean for obtain the robust estimation}

\item{conf:}{a value between 0 and 1 with the desired probability in the
credible intervals}
}
\value{
a data frame with all the important fitted parameters
}
\description{
Summary of an arima(p,d,q) model  in STAN
}
\details{
The function returns a data.frame object with the fitted parameters
}
\author{
Asael Alonzo Matamoros
}
