% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sunspots.R
\docType{data}
\name{sunspots}
\alias{sunspots}
\title{Monthly sunspots numbers 1749-1983}
\format{A time series of monthly data from 1749 to 1983.}
\source{
\pkg{CarData}
}
\usage{
sunspots
}
\description{
**Description**
Monthly mean relative sunspot numbers from 1749 to 1983. Collected at Swiss Federal
Observatory, Zurich until 1960, then Tokyo Astronomical Observatory.
}
\references{
Andrews, D. F. and Herzberg, A. M. (1985) Data: A Collection
of Problems from Many Fields for the Student and Research
Worker. New York: Springer-Verlag.
}
\keyword{datasets}
