% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varstan-package.R
\docType{package}
\name{varstan-package}
\alias{varstan-package}
\alias{varstan}
\title{The 'varstan' package.}
\description{
The goal of this package is analyze time series with structured models such as arima, m-garch, Varma, m-Bekk, SVD, HMM and state space models. So far a arima, garch and varma models are aviable for use.
}
\references{
Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. https://mc-stan.org
}
