% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{vector_to_matrix}
\alias{vector_to_matrix}
\title{Extract lags from vector}
\usage{
extractm(indx,fit,n,d)
}
\arguments{
\item{vect}{a vector with the posterior estimate}

\item{d}{an integer with the number of columns}

\item{p}{the number of lags to be extracted}
}
\value{
a  list of matrices with the extracted lags
}
\description{
Extract a Matrix from an stan fit summary
}
\details{
The function returns a matrix with the extracted coefficients of the stan fit
summary
}
\author{
Asael Alonzo Matamoros
}
