% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Others.R
\name{version}
\alias{version}
\title{print the current available models in the varstan package version}
\usage{
version(View = FALSE)
}
\arguments{
\item{View}{a boolean value, by default is False, if \code{View} is TRUE then
a data frame with the current models will visualize in the RStudio console}
}
\value{
a string
}
\description{
The functions prints the name,version, algorithm and current models implemented in
the varstan package
}
\examples{
 version()
 parameters(classes = "Sarima")
 distribution(par = "ar")

}
\author{
Asael Alonzo Matamoros.
}
