% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Astrovan.R
\docType{data}
\name{Astrovan}
\alias{Astrovan}
\title{Random Multivariate time series Data}
\format{A matrix with of 100 rows and 2 columns
\describe{
 \item{Y1}{First time series}
 \item{Y2}{Second time series}
}}
\source{
Randomly generated data
}
\usage{
Astrovan
}
\description{
A dataset containing randomly generated multivariate(d = 2) time series data that follows a var(1)-bekk(1,1)-m(1) model
as follows. the var proces is a 0.1 matrix, the arch process has alha1 = 0.1(I)  and alpha1 = 0.3(I), the garch process
has a beta1 = 0.2*I and the mgarch has a H = 0.15 matrix. Where I is the identity matrix.
}
\keyword{datasets}
