% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.varstan}
\alias{autoplot.varstan}
\title{autoplot methods for varstan models}
\usage{
\method{autoplot}{varstan}(object, par = "fit", prob = 0.9, combo = c("dens", "trace"), ...)
}
\arguments{
\item{object}{An object of class \code{varstan}.}

\item{par}{a character string with the desired plot. Valid values are \code{"fit"} for the
model's fitted values plot, \code{"residuals"} for the residuals posterior mean plot,
and \code{"parameters"} for density and trace plots for the sampled model parameter.
The default value is \code{"fit"}.}

\item{prob}{A number \eqn{p \in (0,1)}{p (0 < p < 1)} indicating the desired
probability mass to include in the intervals. The default is to report
\eqn{90}\% intervals (\code{prob=0.9}) rather than the traditionally used
\eqn{95}\%.}

\item{combo}{A character vector with at least two elements. Each element of combo corresponds
to a column in the resulting graphic and should be the name of one of the available MCMC functions.}

\item{...}{Further arguments passed to  \code{mcmc_combo}.}
}
\value{
An autoplot object from ggplot2 class.
}
\description{
Preliminar autoplot methods for varstan models only valid for univariate time series models.
The function prints the fitted values time series, the trace and density plots for the
sampled model parameters, or the residuals' posterior mean time series.
}
\examples{
\dontrun{
 sf1 = auto.sarima(ts = birth)
 # fitted model
 autoplot(sf1)

 # residuals
 autoplot(sf1,par = "residuals")

 # parameters
 autoplot(sf1,par = "parameter")
}

}
