% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bekk.R
\name{Bekk}
\alias{Bekk}
\title{Constructor varma(p,q)-Bekk(s,k) object}
\usage{
Bekk(ts,order = c(1,1,0),varma = c(0,0),genT = FALSE)
}
\arguments{
\item{ts}{an multivariate time series}

\item{order}{an optional value for specify the Bekk  part for variance, same as
order parameter:  the three components (s, k, h) are the arch order,
the garch order and the garch on the mean (mgarch) order.}

\item{varma}{A specification of VARMA model: the two components (p, q) are the AR
order, the number of differences, and the MA order.}

\item{genT}{a boolean value to specify a Generalized a t-student model Cruz (2015)}

\item{series.name}{an optional string vector with the series names.}
}
\description{
Constructor of the varma(p,q)-Bekk(s,k) object for Bayesian estimation in STAN
}
\details{
The function returns  a list with the data for running stan() function of
 rstan package

If \code{sd} option can only be used with the mbekk function, and it  adds
a m-Bekk model for volatility.

If a mbekk model is not used, then sigma0 represents the covariance matrix
of the model.

If a bekk model is specified, then sigma0 represents the triangular inferior matrix
alpha0 of a bekk model

The default priors used in varma are:

\itemize{
 \item{"ar"}{ar ~ normal(0,0.5)}
 \item{"ma"}{ma ~ normal(0,0.5)}
 \item{"mu0"}{mu0 ~ normal(0,1)}
 \item{"sigma0"}{sigma0 ~ t-student(0,1,7)}
 \item{"arch"}{arch ~ normal(0,0.5)}
 \item{"garch"}{garch ~ normal(0,0.5)}
 \item{"mgarch"}{mgarch ~ normal(0,0.5)}
 \item{"dfv"}{dfv ~ gamma(2,0.1)}
}

For changing the default prior use the function \code{set_prior}
}
\examples{
# Declare a Bekk model for the Astrovan data
model = Bekk(Astrovan,order = c(1,1,0),varma = c(1,0))
model


}
\references{
Polasek, Ren(1999).
 A multivariate GARCH-M model for exchange rates in the US,Germany and Japan

 Fonseca,Ferreira, Migon (2008)
 Objective Bayesian analysis for the Student-t regression model
}
\seealso{
\code{\link{garch}} \code{\link{Sarima}} \code{\link{set_prior}}
}
\author{
Asael Alonzo Matamoros
}
