% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garch.R
\name{garch}
\alias{garch}
\title{A  constructor for a garch(s,k,h) model}
\usage{
garch(ts,s,k)
}
\arguments{
\item{ts}{an univariateivariate time series}

\item{s}{an integer with the order of the arch(s) part}

\item{k}{an integer with the order of the garch(k) part}

\item{h}{an integer with the order of the mean part}
}
\value{
a list with the components
\itemize{
 \item n: the length of the time series
 \item s: an integer with the order of the arch coefficients
 \item k: an integer with the order of the garch coefficients
 \item h: an integer with the order of the mgarch coefficients
 \item y: vector with the multivariate time series
 \item prior_arch:   a matrix with the hyper-parameters for the arch coefficients
 \item prior_garch:  a matrix with the hyper-parameters for the garch  coefficients
 \item prior_mgarch: a matrix with the hyper-parameters for the mgarch  coefficients
 \item p: an integer with the order of ar coefficients
 \item q: an inter withe the order of ma coefficients
 \item prior_arch:   a matrix with the hyper-parameters for the arch coefficients
}
}
\description{
Constructor of the garch(s,k,h) object for bayesian estimation in STAN
}
\details{
The function returns  a list with the data for running stan() function of
 rstan package
}
\seealso{
\code{\link{arima}}
}
\author{
Asael Alonzo Matamoros
}
