% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garch.R
\name{garch}
\alias{garch}
\title{A  constructor for a garch(s,k,h) model}
\usage{
garch(ts,order = (1,1,0),arma = c(0,0),genT = FALSE)
}
\arguments{
\item{ts}{an univariate time series}

\item{order}{A specification of the garch  model: the three components (s, k, h)
are the arch order, the garch order, and the mgarch order.}

\item{arma}{A specification of the  ARMA model,same as order parameter:  the two
components (p, q) are the AR order,and the  MA order.}

\item{xreg}{Optionally, a numerical matrix of external regressors,
which must have the same number of rows as ts. It should not be a data frame.}

\item{genT}{a boolean value to specify for a generalized t-student garch model}

\item{series.name}{an optional string vector with the series names.}
}
\description{
Constructor of the garch(s,k,h) object for Bayesian estimation in STAN
}
\details{
The function returns a list with the data for running stan() function of
 rstan package

The default priors used in Sarima are:

\itemize{
 \item{ar ~ normal(0,0.5)}
 \item{ma ~ normal(0,0.5)}
 \item{mu0 ~ t-student(0,2.5,6)}
 \item{sigma0 ~ t-student(0,1,7)}
 \item{arch ~ normal(0,0.5)}
 \item{garch ~ normal(0,0.5)}
 \item{mgarch ~ normal(0,0.5)}
 \item{dfv ~ gamma(2,0.1)}
 \item{breg ~ t-student(0,2.5,6)}
}

For changing the default prior use the function \code{set_prior}
}
\examples{
\dontrun{
dat = garch(birth,order = c(1,1,0),arma = c(1,1))
dat
}

}
\references{
Engle, Robert F. (1982). ?Autoregressive Conditional Heteroscedasticity with
  Estimates of the Variance of United Kingdom Inflation.
  Econometrica 50 (4): 987-1007. JSTOR 1912773.
}
\seealso{
\code{\link{Sarima}} \code{\link{auto.arima}} \code{\link{set_prior}}
}
\author{
Asael Alonzo Matamoros
}
