% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garch.R
\name{garch}
\alias{garch}
\title{A  constructor for a garch(s,k,h) model}
\usage{
garch(ts,s,k)
}
\arguments{
\item{ts}{an univariateivariate time series}

\item{genT}{a boolean value to specify for a generalized t-student garch model


The default priors used in Sarima are:

\itemize{
 \item{"ar"}{ar ~ normal(0,0.5)}
 \item{"ma"}{ma ~ normal(0,0.5)}
 \item{"mu0"}{mu0 ~ t-student(0,2.5,6)}
 \item{"sigma0"}{sigma0 ~ t-student(0,1,7)}
 \item{"arch"}{arch ~ normal(0,0.5)}
 \item{"garch"}{garch ~ normal(0,0.5)}
 \item{"mgarch"}{mgarch ~ normal(0,0.5)}
 \item{"dfv"}{dfv ~ gamm(2,0.1)}
}

For changing the default prior use the function \code{set_prior}}

\item{order:}{A specification of thegarch  model: thethree components (s, k, h)
are the arch order, the garch order, and the mgarch order.}

\item{arma:}{A specification of the  ARMA model,same as order parameter:  the two
components (p, q) are the AR order,and the  MA order.}
}
\description{
Constructor of the garch(s,k,h) object for bayesian estimation in STAN
}
\details{
The function returns  a list with the data for running stan() function of
 rstan package
}
\references{
Engle, Robert F. (1982). ?Autoregressive Conditional Heteroscedasticity with
  Estimates of the Variance of United Kingdom Inflation?.
  Econometrica 50 (4): 987-1007. JSTOR 1912773.
}
\seealso{
\code{\link{Sarima}} \code{\link{auto.arima}} \code{\link{set_prior}}
}
\author{
Asael Alonzo Matamoros
}
