% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varstan.R
\name{get_df}
\alias{get_df}
\title{Get the degree freedom values of a varma model}
\usage{
get_df(fit,robust = FALSE,...)
}
\arguments{
\item{obj}{a varstan object}

\item{robust}{a boolean for obtain the robust estimation}
}
\value{
a data frame with the degree freedom time series parameter
}
\description{
get the degree freedom values of a varma(p,q) model  in STAN
}
\details{
The function returns a data.frame object with the degree freedom values
}
\author{
Asael Alonzo Matamoros
}
