% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{get_prior}
\alias{get_prior}
\title{Generic function for setting a prior to an specify parameter}
\usage{
set_prior(type,lag,par1,par2,dist)
}
\arguments{
\item{dat:}{an arima model speficied in varstan.}

\item{type:}{the type of parameter wich a prior is defined could be mu, sigma0, ar, ma, arch, garch, mgarch}

\item{lag:}{the lag of the parameter which the prior is defined if lag = 0, then the prior will be applied
for all the lags.}
}
\description{
Generic function for setting a prior to an specify parameter
}
\examples{

dat = Sarima(ipc,order = c(1,1,2))
dat = set_prior(dat,type = "ar",par1 =0,par2 = 1,dist = "normal")
get_prior(dat,type = "ar")

dat = set_prior(dat,type = "mu0",par1 = 0,par2 = 2.5,df = 7,dist = "student")
get_prior(dat,type = "mu0")

dat = set_prior(dat,type = "ma",par1 =2,par2 = 2,dist = "beta",lag = 2)
get_prior(dat,type = "ma")

}
\author{
Asael Alonzo Matamoros
}
