% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varstan.R
\name{get_rstan}
\alias{get_rstan}
\title{Extract a stanfit object implemented in rstan package}
\usage{
get_rstan(obj)
}
\arguments{
\item{obj:}{a varstan object}
}
\value{
a stanfit object
}
\description{
Extract a stanfit object implemented in rstan package
}
\author{
Asael Alonzo Matamoros
}
