% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{point_estimate}
\alias{point_estimate}
\title{Set the generic function for point_estimate a varstan class}
\usage{
point_estimate(model, fit, robust, conf, ...)
}
\description{
A point estimate of a desired model
}
\details{
@method point_estimate

@export
}
