% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_fit.R
\name{posterior_fit}
\alias{posterior_fit}
\title{Generic function and method for extract the fitted values}
\usage{
posterior_fit(obj)
}
\arguments{
\item{obj:}{A varstan object, \code{\link[=varstan]{varstan}}}

\item{robust:}{A boolean value, if its \code{TRUE} it returns the median of the posterior distribution,
and if its \code{FALSE} it returns the mean, by default is the \code{FALSE} value}
}
\value{
An array with the fitted values of the time series model, if a varma model is used, then
the return will be a matrix with columns as the dimension of the time series, and rows as the length
}
\description{
The function returns the posterior estimate of the fitted values
of a varstan model, similar to the fit_values functions of other
packages.
}
\author{
Asael Alonzo Matamoros
}
