% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sarima.R
\name{Sarima}
\alias{Sarima}
\title{Constructor Seasonal arima model}
\usage{
Sarima(ts,order = c(0,0,0),seasonal = c(0,0,0),xreg = NULL,period = 0)
}
\arguments{
\item{ts}{a univariate time series}

\item{order:}{A specification of the non-seasonal part of the ARIMA model: the
three components (p, d, q) are the AR order, the degree of differencing, and the
MA order.}

\item{seasonal:}{A specification of the seasonal part of the ARIMA model,same as
order parameter:  the three components (p, d, q) are the seasonal AR order,
the degree of seasonal differencing, and the seasonal MA order.}

\item{xreg:}{Optionally, a numerical matrix of external regressors,
which must have the same number of rows as ts. It should not be a data frame.}

\item{period:}{an integer specifying the periodicity of the time series by
deafault the value frequency(ts) is used.}
}
\description{
Constructor of the Sarima model  for bayesian estimation in STAN
}
\details{
The function returns  a list with  all the neccesary data for running a
Seasonal arima model in stan

If \code{xreg} option is used, the model by default will cancel the
seasonal differences adjusted (D = 0). If a value \code{d} > 0 is used, all
the regressor variables in xreg will be differenced aswell

The default priors used in Sarima are:

\itemize{
 \item{"ar"}{ar ~ normal(0,0.5)}
 \item{"ma"}{ma ~ normal(0,0.5)}
 \item{"mu0"}{mu0 ~ t-student(0,2.5,6)}
 \item{"sigma0"}{sigma0 ~ t-student(0,1,7)}
 \item{"seasonal ar"}{sar ~ normal(0,0.5)}
 \item{"seasonal ma"}{sma ~ normal(0,0.5)}
 \item{"seasonal breg"}{breg ~ t-student(0,2.5,6)}
}

For changing the default prior use the function \code{set_prior}
}
\examples{
# Declare a seasonal arima model for the birth data of asta package

model = Sarima(bth,order = c(0,1,2),seasonal = c(1,1,1))
model


}
\references{
Box, George; Jenkins, Gwilym (1970).
 Time Series Analysis: Forecasting and Control. San Francisco: Holden-Day.
}
\seealso{
\code{\link{garch}} \code{\link{auto.arima}} \code{\link{set_prior}}
}
\author{
Asael Alonzo Matamoros
}
