% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{set_prior}
\alias{set_prior}
\title{Prior objects for time series models}
\usage{
set_prior(type,lag,par1,par2,dist)
}
\arguments{
\item{dat}{an arima model specified in varstan.}

\item{type}{the type of parameter which a prior is defined could be mu,
sigma0, ar, ma, arch, garch, mgarch}

\item{lag}{an optional integer value indicated thedesired lag of the parameter which the prior
is defined if lag = 0, then the prior will be applied for all the lags.}

\item{par1}{The first hyper-parameter of the prior distribution}

\item{par2}{The second hyper-parameter of the prior distribution}

\item{dist}{the distribution of the prior distribution}

\item{df}{The parameter for the degree freedom only for t-student and
chi-square}
}
\description{
Generic function for setting a prior to an specify parameter
}
\examples{

dat = Sarima(birth,order = c(1,1,2))
dat = set_prior(dat,type = "ar",par1 =0,par2 = 1,dist = "normal")
get_prior(dat,type = "ar")

dat = set_prior(dat,type = "mu0",par1 = 0,par2 = 2.5,df = 7,dist = "student")
get_prior(dat,type = "mu0")

dat = set_prior(dat,type = "ma",par1 =2,par2 = 2,dist = "beta",lag = 2)
get_prior(dat,type = "ma")

}
\author{
Asael Alonzo Matamoros
}
