% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varma.R
\name{varma}
\alias{varma}
\title{Constructor var(p)-bekk(s,k) object}
\usage{
varbekk(ts,p,s,k)
}
\arguments{
\item{ts}{an multivariate time series}

\item{p}{an integer with the order of the var(p) part}

\item{genT}{a boolean value to specify a Generalized a t-student model Cruz (2015)}

\item{s}{an integer with the order of the arch(s) part}

\item{k}{an integer with the order of the arch(k) part}
}
\value{
a list with the components
\itemize{
 \item n: the length of the time series
 \item d: the dimension of the time series
 \item p: an integer with the order of the var coefficients
 \item s: an integer with the order of the arch coefficients
 \item k: an interger with the order of th garch coefficients
 \item v: a real with degree freedom for a t-student sample
 \item y: vector with the multivariate time series
 \item prior_var:   a vector with the hyper-parameters for the var   coefficient
 \item prior_arch:  a vector with the hyper-parameters for the arch  coefficient
 \item prior_garch: a vector with the hyper-parameters for the garch coefficient
}
}
\description{
Constructor of the var(p)-bekk(s,k) object for bayesian estimation in STAN
}
\details{
The function returns  a list with the data for running stan() function of
 rstan package
}
\author{
Asael Alonzo Matamoros
}
