% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varstan.R
\name{varstan}
\alias{varstan}
\title{Constructor varstan) object}
\usage{
varstan(model)
}
\arguments{
\item{model}{A time series object for the varstan models}

\item{chains}{the number of chains to be run}

\item{iter}{the number of iteration per chain}

\item{warmup}{the number of initial iteration to be burned}

\item{adapt.delta}{the thin of the jumps in a HMC method}
}
\value{
a list with the components
\itemize{
 \item stanfit: An stanfit object returned by rstan
 \item model: the current fitted model
 \item stanparams: The stan parameters used in the HMC NUTS algorithm
}
}
\description{
Constructor of the varstan object for bayesian estimation in STAN
}
\details{
The function returns  a list with the data for running stan() function of
 rstan package
}
\author{
Asael Alonzo Matamoros
}
